requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
langgraph>=0.4.0
openai>=1.0.0
tiktoken>=0.9.0
chromadb>=1.0.0
fakeredis>=2.0.0
redis>=5.0.0
python-dotenv>=1.0.0